Added the ability to read Version 4 .anm files.
It's not fully tested and may be a little buggy but it's a start. However, they still won't work in the program.  I don't know how to map the bone ID's in the .anm file to the bone ID's in the .skl.  I think the unknown data section in the newest .skl files contains the mapping.